# obsrvr-lake/stellar-query-api/go/docs/swagger.json
{
    "basePath": "/",
    "definitions": {
...
        "main.LatestPrice": {
            "properties": {
                "base_asset": {
                    "type": "string"
                },
                "counter_asset": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "trade_count_24h": {
                    "type": "integer"
                },
                "volume_24h": {
                    "type": "number"
                }
            },
            "type": "object"
        },
...
    },
    "host": "gateway.withobsrvr.com",
    "info": {
        "contact": {
            "email": "support@obsrvr.com",
            "name": "OBSRVR Team",
            "url": "https://obsrvr.com"
        },
        "description": "API for querying Stellar blockchain data across Bronze, Silver, Semantic, and Gold data layers.\n\n## Data Layers\n- **Bronze**: Raw blockchain data (ledgers, transactions, operations, effects)\n- **Silver**: Enriched and processed data (accounts, assets, transfers, trades)\n- **Semantic**: Meaning-oriented views (activities, contracts, value flows)\n- **Gold**: Analytical views (snapshots, compliance reports)\n- **Index Plane**: Fast lookup services (transaction hash, contract events)\n\n## Amount Formatting\nAll amounts are returned in decimal format with 7 decimal places (Stellar stroops conversion).\nExample: Raw stroops `10000000` is formatted as `\"1.0000000\"`",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "Stellar Query API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/silver/events/contract/{contract_id}": {
            "get": {
                "description": "Returns raw contract events for a specific contract from bronze layer",
                "parameters": [
                    {
                        "description": "Contract ID (hex format)",
                        "in": "path",
                        "name": "contract_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Filter by event type",
                        "in": "query",
                        "name": "event_type",
                        "type": "string"
                    },
                    {
                        "default": 20,
                        "description": "Max results (default: 20, max: 200)",
                        "in": "query",
                        "name": "limit",
                        "type": "integer"
                    },
                    {
                        "description": "Pagination cursor",
                        "in": "query",
                        "name": "cursor",
                        "type": "string"
                    },
                    {
                        "default": "desc",
                        "description": "Sort order",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "in": "query",
                        "name": "order",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Contract events with pagination",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing contract_id",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get contract events by contract ID",
                "tags": [
                    "Events"
                ]
            }
        },
        "/api/v1/silver/events/generic": {
            "get": {
                "description": "Returns raw contract events from bronze layer with optional filters",
                "parameters": [
                    {
                        "description": "Filter by contract ID (C...)",
                        "in": "query",
                        "name": "contract_id",
                        "type": "string"
                    },
                    {
                        "description": "Filter by event type: contract, system, diagnostic",
                        "in": "query",
                        "name": "event_type",
                        "type": "string"
                    },
                    {
                        "description": "Search within topics_decoded",
                        "in": "query",
                        "name": "topic_match",
                        "type": "string"
                    },
                    {
                        "description": "Start of ledger range",
                        "in": "query",
                        "name": "start_ledger",
                        "type": "integer"
                    },
                    {
                        "description": "End of ledger range",
                        "in": "query",
                        "name": "end_ledger",
                        "type": "integer"
                    },
                    {
                        "default": 20,
                        "description": "Max results (default: 20, max: 200)",
                        "in": "query",
                        "name": "limit",
                        "type": "integer"
                    },
                    {
                        "description": "Pagination cursor",
                        "in": "query",
                        "name": "cursor",
                        "type": "string"
                    },
                    {
                        "default": "desc",
                        "description": "Sort order",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "in": "query",
                        "name": "order",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Generic events with pagination",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get generic contract events",
                "tags": [
                    "Events"
                ]
            }
        },
...
        "/api/v1/silver/prices/{base}/{counter}/latest": {
            "get": {
                "description": "Returns the most recent trade price for a trading pair with 24h volume",
                "parameters": [
                    {
                        "description": "Base asset code",
                        "in": "path",
                        "name": "base",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Counter asset code",
                        "in": "path",
                        "name": "counter",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Base asset issuer",
                        "in": "query",
                        "name": "base_issuer",
                        "type": "string"
                    },
                    {
                        "description": "Counter asset issuer",
                        "in": "query",
                        "name": "counter_issuer",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Latest price",
                        "schema": {
                            "$ref": "#/definitions/main.LatestPrice"
                        }
                    },
                    "404": {
                        "description": "No trades found",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get latest price",
                "tags": [
                    "Prices"
                ]
            }
        },
        "/api/v1/silver/prices/{base}/{counter}/ohlc": {
            "get": {
                "description": "Returns OHLC (Open/High/Low/Close) candles for a trading pair, aggregated by time interval",
                "parameters": [
                    {
                        "description": "Base asset code (e.g., XLM, USDC)",
                        "in": "path",
                        "name": "base",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Counter asset code",
                        "in": "path",
                        "name": "counter",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": "1h",
                        "description": "Time interval: 1m, 5m, 15m, 1h, 4h, 1d, 1w",
                        "in": "query",
                        "name": "interval",
                        "type": "string"
                    },
                    {
                        "description": "Base asset issuer (omit for XLM)",
                        "in": "query",
                        "name": "base_issuer",
                        "type": "string"
                    },
                    {
                        "description": "Counter asset issuer (omit for XLM)",
                        "in": "query",
                        "name": "counter_issuer",
                        "type": "string"
                    },
                    {
                        "default": "24 hours ago",
                        "description": "Start time (RFC3339)",
                        "in": "query",
                        "name": "start",
                        "type": "string"
                    },
                    {
                        "default": "now",
                        "description": "End time (RFC3339)",
                        "in": "query",
                        "name": "end",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OHLC candles",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get OHLC price candles",
                "tags": [
                    "Prices"
                ]
            }
        },
        "/api/v1/silver/prices/pairs": {
            "get": {
                "description": "Returns available trading pairs with trade counts (minimum 5 trades)",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Trading pairs",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "List trading pairs",
                "tags": [
                    "Prices"
                ]
            }
        },
...
    },
    "schemes": [
        "https"
    ],
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Format: \"Api-Key YOUR_API_KEY\"",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}